archive/voice-analysis/requirements.txt
faster-whisper>=1.0.0
scipy>=1.11.0
numpy>=1.24.0
pydub>=0.25.1
PyYAML>=6.0
anthropic>=0.39.0
httpx>=0.27.0
